+++ StudentPerfomance-Backend/SPerfomance.Application.SemesterPlans.Module/Commands/Delete/DeleteCommand.cs
@@ -6,9 +6,11 @@ using SPerfomance.Application.Shared.Module.CQRS.Commands;
 using SPerfomance.Application.Shared.Module.Operations;
 using SPerfomance.Application.Shared.Module.Schemas.SemesterPlans;
 using SPerfomance.Application.Shared.Module.Schemas.Semesters;
+using SPerfomance.Application.Shared.Users.Module.Commands.Common;
 using SPerfomance.Domain.Module.Shared.Common.Abstractions.Repositories;
 using SPerfomance.Domain.Module.Shared.Entities.SemesterPlans;
 using SPerfomance.Domain.Module.Shared.Entities.Semesters;
+using SPerfomance.Domain.Module.Shared.Entities.Users;
 
 namespace SPerfomance.Application.SemesterPlans.Module.Commands.Delete;
 
@@ -20,20 +22,33 @@ internal sealed class DeleteCommand : ICommand
 
 	public readonly ICommandHandler<DeleteCommand, SemesterPlan> Handler;
 
-	public DeleteCommand(SemesterSchema semester, SemesterPlanSchema plan)
+	public DeleteCommand(SemesterSchema semester, SemesterPlanSchema plan, string token)
 	{
 		_getSemester = ExpressionsFactory.GetSemester(semester.ToRepositoryObject());
 		_plan = plan;
 		_repository = new SemesterPlansCommandRepository();
-		Handler = new CommandHandler(_repository);
+		Handler = new VerificationHandler<DeleteCommand, SemesterPlan>(token, User.Admin);
+		Handler = new CommandHandler(Handler, _repository);
 	}
 
-	internal sealed class CommandHandler(SemesterPlansCommandRepository repository) : ICommandHandler<DeleteCommand, SemesterPlan>
+	internal sealed class CommandHandler : DecoratedCommandHandler<DeleteCommand, SemesterPlan>
 	{
-		private readonly SemesterPlansCommandRepository _repository = repository;
+		private readonly SemesterPlansCommandRepository _repository;
 
-		public async Task<OperationResult<SemesterPlan>> Handle(DeleteCommand command)
+		public CommandHandler(
+			ICommandHandler<DeleteCommand, SemesterPlan> handler,
+			SemesterPlansCommandRepository repository
+			) : base(handler)
 		{
+			_repository = repository;
+		}
+
+		public override async Task<OperationResult<SemesterPlan>> Handle(DeleteCommand command)
+		{
+			OperationResult<SemesterPlan> result = await base.Handle(command);
+			if (result.IsFailed)
+				return result;
+
 			Result<SemesterPlan> delete = await _repository.Remove(command._getSemester, command._plan.Discipline);
 			return delete.IsFailure ?
 				new OperationResult<SemesterPlan>(delete.Error) :
