+++ StudentPerfomance-Backend/SPerfomance.Application.Students.Module/Commands/Remove/RemoveCommand.cs
@@ -4,11 +4,13 @@ using SPerfomance.Application.Shared.Module.CQRS.Commands;
 using SPerfomance.Application.Shared.Module.Operations;
 using SPerfomance.Application.Shared.Module.Schemas.StudentGroups;
 using SPerfomance.Application.Shared.Module.Schemas.Students;
+using SPerfomance.Application.Shared.Users.Module.Commands.Common;
 using SPerfomance.Application.Students.Module.Repository;
 using SPerfomance.Application.Students.Module.Repository.Expressions;
 using SPerfomance.Domain.Module.Shared.Common.Abstractions.Repositories;
 using SPerfomance.Domain.Module.Shared.Entities.StudentGroups;
 using SPerfomance.Domain.Module.Shared.Entities.Students;
+using SPerfomance.Domain.Module.Shared.Entities.Users;
 
 namespace SPerfomance.Application.Students.Module.Commands.Remove;
 
@@ -17,21 +19,35 @@ internal sealed class RemoveCommand : ICommand
 	private readonly IRepositoryExpression<Student> _getStudent;
 	private readonly IRepositoryExpression<StudentGroup> _getGroup;
 	private readonly StudentCommandRepository _repository;
+
 	public readonly ICommandHandler<RemoveCommand, Student> Handler;
-	public RemoveCommand(StudentSchema student)
+
+	public RemoveCommand(StudentSchema student, string token)
 	{
 		_getStudent = ExpressionsFactory.GetStudent(student.ToRepositoryObject());
 		_getGroup = ExpressionsFactory.GetGroupByName(student.ToRepositoryObject());
 		_repository = new StudentCommandRepository();
-		Handler = new CommandHandler(_repository);
+		Handler = new VerificationHandler<RemoveCommand, Student>(token, User.Admin);
+		Handler = new CommandHandler(Handler, _repository);
 	}
 
-	internal sealed class CommandHandler(StudentCommandRepository repository) : ICommandHandler<RemoveCommand, Student>
+	internal sealed class CommandHandler : DecoratedCommandHandler<RemoveCommand, Student>
 	{
-		private readonly StudentCommandRepository _repository = repository;
+		private readonly StudentCommandRepository _repository;
 
-		public async Task<OperationResult<Student>> Handle(RemoveCommand command)
+		public CommandHandler(
+			ICommandHandler<RemoveCommand, Student> handler,
+			StudentCommandRepository repository) : base(handler)
 		{
+			_repository = repository;
+		}
+
+		public override async Task<OperationResult<Student>> Handle(RemoveCommand command)
+		{
+			var result = await base.Handle(command);
+			if (result.IsFailed)
+				return result;
+
 			Result<Student> delete = await _repository.Remove(command._getStudent, command._getGroup);
 			return delete.IsFailure ?
 				new OperationResult<Student>(delete.Error) :
