+++ StudentPerfomance-Backend/SPerfomance.Application.TeacherDepartments.Module/Commands/Delete/TeacherDepartmentDeleteCommand.cs
@@ -3,10 +3,12 @@ using CSharpFunctionalExtensions;
 using SPerfomance.Application.Shared.Module.CQRS.Commands;
 using SPerfomance.Application.Shared.Module.Operations;
 using SPerfomance.Application.Shared.Module.Schemas.Departments;
+using SPerfomance.Application.Shared.Users.Module.Commands.Common;
 using SPerfomance.Application.TeacherDepartments.Module.Repository;
 using SPerfomance.Application.TeacherDepartments.Module.Repository.Expressions;
 using SPerfomance.Domain.Module.Shared.Common.Abstractions.Repositories;
 using SPerfomance.Domain.Module.Shared.Entities.TeacherDepartments;
+using SPerfomance.Domain.Module.Shared.Entities.Users;
 
 namespace SPerfomance.Application.TeacherDepartments.Module.Commands.Delete;
 
@@ -17,19 +19,31 @@ internal sealed class TeacherDepartmentDeleteCommand : ICommand
 
 	public readonly ICommandHandler<TeacherDepartmentDeleteCommand, TeachersDepartment> Handler;
 
-	public TeacherDepartmentDeleteCommand(DepartmentSchema department)
+	public TeacherDepartmentDeleteCommand(DepartmentSchema department, string token)
 	{
 		_expression = ExpressionsFactory.GetDepartment(department.ToRepositoryObject());
 		_repository = new TeacherDepartmentsCommandRepository();
-		Handler = new CommandHandler(_repository);
+		Handler = new VerificationHandler<TeacherDepartmentDeleteCommand, TeachersDepartment>(token, User.Admin);
+		Handler = new CommandHandler(Handler, _repository);
 	}
 
-	internal sealed class CommandHandler(TeacherDepartmentsCommandRepository repository) : ICommandHandler<TeacherDepartmentDeleteCommand, TeachersDepartment>
+	internal sealed class CommandHandler : DecoratedCommandHandler<TeacherDepartmentDeleteCommand, TeachersDepartment>
 	{
-		private readonly TeacherDepartmentsCommandRepository _repository = repository;
+		private readonly TeacherDepartmentsCommandRepository _repository;
 
-		public async Task<OperationResult<TeachersDepartment>> Handle(TeacherDepartmentDeleteCommand command)
+		public CommandHandler(
+			ICommandHandler<TeacherDepartmentDeleteCommand, TeachersDepartment> handler,
+			TeacherDepartmentsCommandRepository repository) : base(handler)
 		{
+			_repository = repository;
+		}
+
+		public override async Task<OperationResult<TeachersDepartment>> Handle(TeacherDepartmentDeleteCommand command)
+		{
+			var result = await base.Handle(command);
+			if (result.IsFailed)
+				return result;
+
 			Result<TeachersDepartment> delete = await _repository.Remove(command._expression);
 			return delete.IsFailure ?
 				new OperationResult<TeachersDepartment>(delete.Error) :
